+++ src/components/UserInput.vue
@@ -1,16 +1,20 @@
 <template>
-    <q-input dark dense filled autofocus autogrow :placeholder="t('userInput.placeholder')" color="white"
-        style="min-width: 85%; max-width: 85%;" @keydown.enter.prevent="handleUserInput" v-model="question">
+    <q-input dense filled autofocus autogrow style="min-width: 70%; max-width: 70%;" 
+        dark color="white" bg-color="grey-9"
+        :placeholder="t('userInput.placeholder')" @keydown.enter.prevent="handleUserInput" v-model="question">
+
         <template v-slot:prepend>
             <q-icon name="live_help" style="padding: 5px" color="white" />
         </template>
+
         <template v-slot:append>
             <q-btn @click="handleUserInput" dense flat icon="send" color="white">
-                <q-tooltip delay="750" transition-show="scale" transition-hide="scale">
+                <q-tooltip :delay="750" transition-show="scale" transition-hide="scale">
                     {{ t('userInput.tooltip.send') }}
                 </q-tooltip>
             </q-btn>
         </template>
+
     </q-input>
 </template>